bug fixes
- fix old object redirector issues.
- removed references to old quizzes.
- added Book and Author into combobox so that user knows exactly where they are adding questions.
- removed UE default data structure.
