--- Content/DanielStuff/reference
@@ -1,23 +1,0 @@
-radiobox
-https://www.youtube.com/watch?v=gULcJQwagmI
-
-show mouse cursor during game at level blueprint
-https://www.youtube.com/watch?v=SA4KZFJVL9E
-
-create a interaction when near an object
-https://www.youtube.com/watch?v=tob3GEdMJFc
-
-how to pause a game when menu is opened
-https://www.youtube.com/watch?v=Yr8beK9FMe0
-
-UIwidget
-https://www.youtube.com/watch?v=hv8cClvSvSo
-
-UIcreation
-https://www.youtube.com/watch?v=SRsRrYZGCgQ
-
-SQL - no step-by-step
-https://www.youtube.com/watch?v=8Nsqz_AMOCI
-
-Quiz2
-https://www.youtube.com/watch?v=vln3KR9Aqxw
