+++ Saved/Config/CrashReportClient/UE4CC-Windows-147C557B4FD893E82C2D528828336C1D/CrashReportClient.ini
@@ -1,0 +1,5 @@
+[CrashReportClient]
+bHideLogFilesOption=false
+bIsAllowedToCloseWithoutSending=true
+CrashConfigPurgeDays=2
+
