+++ Saved/Config/CrashReportClient/UE4CC-Windows-760E96B34D1733453C9B0C80A8549D56/CrashReportClient.ini
@@ -1,0 +1,5 @@
+[CrashReportClient]
+bHideLogFilesOption=false
+bIsAllowedToCloseWithoutSending=true
+CrashConfigPurgeDays=2
+
