+++ Saved/Config/Windows/Engine.ini
@@ -57,11 +57,3 @@ TunnelMulticastEndpoint=
 [/Script/LuminPlatformEditor.MagicLeapSDKSettings]
 MLSDKPath=(Path="")
 
-[WindowsApplication.Accessibility]
-StickyKeysHotkey=True
-ToggleKeysHotkey=True
-FilterKeysHotkey=True
-StickyKeysConfirmation=True
-ToggleKeysConfirmation=True
-FilterKeysConfirmation=True
-
